Cross Fingering fix-ish
Reformatted code to be nested if statements to make it easier to read and manage each case, rather than using flags.
Note debouncing now waits for the note to be steady for 30mS, rather than waiting every time there's a new note to check again.  30mS is borderline not long enough, but it would make fast grace notes or trills not as effective.